backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client>=1.14.2",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "openai-agents>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "supabase>=2.0.0",
    "pydantic[email]>=2.0.0",
]
